Refactor out the various sections of the Room page into components so that they don't re-render each other.
