+++ src/Blazor.Gitter.Core/Components/Pages/Room.razor
@@ -3,47 +3,10 @@
 
 @inherits RoomModel
 
-<div class="blg-top-center d-flex justify-content-start align-items-center text-muted">
-	<img src="@ThisRoom.AvatarUrl" width="48" height="48" />
-	<h3 class="h3-responsive mr-2">@ThisRoom.Name</h3>
-	<small class="text-muted text-truncate">@ThisRoom.Topic</small>
-</div>
+<RoomTitle ChatRoom="@ThisRoom" />
 
-<div id="blgmessagelist" class="blg-center scrollable" onscroll="@MessagesScrolled">
-	@if (Messages is object)
-	{
-	<div class="list-group list-group-flush">
-		@foreach (var message in Messages.OrderBy(m => m.Sent))
-		{
-			<Message MessageData="@message" UserId="@State.GetMyUser().Id" />
-		}
+<RoomMessages ChatRoom="@ThisRoom" UserId="@State.GetMyUser().Id" />
 
-	</div>
-	}
-</div>
+<RoomSend ChatRoom="@ThisRoom" User="@State.GetMyUser()" />
 
-<form onsubmit="@SendMessage" class="d-flex align-items-center blg-bottom-center">
-	<img src="@State.GetMyUser().AvatarUrlSmall" width="38" height="38" class="rounded-lg mr-1" />
-	<textarea id="messageinput"
-						class="form-control flex-grow-1"
-						bind="@NewMessage"
-						placeholder="Join the conversation..." />
-	<button type="submit" class="btn btn-primary btn-sm ml-1">Ok</button>
-</form>
-
-<div class="blg-center-right p-1 d-flex flex-column">
-	<h3>Search</h3>
-	<form onsubmit="@Search">
-		<input class="form-control" type="text" placeholder="e.g. context" bind="@SearchText" />
-	</form>
-	<button class="btn btn-secondary form-control" onclick="@(()=>SearchResult=null)">Clear</button>
-	@if (SearchResult is object)
-	{
-		<div class="list-group list-group-flush scrollable">
-			@foreach (var message in SearchResult)
-			{
-				<Message MessageData="@message" UserId="@State.GetMyUser().Id" />
-			}
-		</div>
-	}
-</div>
+<RoomSearch ChatRoom="@ThisRoom" UserId="@State.GetMyUser().Id" />
