+++ src/Blazor.Gitter.Core/Components/Shared/RoomSend.razor
@@ -1,0 +1,10 @@
+﻿@inherits RoomSendBase
+
+<form onsubmit="@SendMessage" class="@OuterClass">
+	<img src="@User.AvatarUrlSmall" width="38" height="38" class="rounded-lg mr-1" />
+	<textarea id="messageinput"
+						class="form-control flex-grow-1"
+						bind="@NewMessage"
+						placeholder="Join the conversation..." />
+	<button type="submit" class="btn btn-primary btn-sm ml-1">Ok</button>
+</form>
