+++ src/Blazor.Gitter.Core/Components/Shared/RoomTitle.razor
@@ -1,0 +1,7 @@
+﻿@inherits RoomTitleBase
+
+<div class="@OuterClass">
+	<img src="@ChatRoom.AvatarUrl" width="48" height="48" />
+	<h3 class="h3-responsive mr-2">@ChatRoom.Name</h3>
+	<small class="text-muted text-truncate">@ChatRoom.Topic</small>
+</div>
